[chore] change calendar package
