+++ app/src/main/java/kr/co/nottodo/view/calendar/monthly/model/CalendarType.kt
@@ -1,0 +1,7 @@
+package kr.co.nottodo.view.calendar.monthly.model
+
+enum class CalendarType {
+    WEEK,
+    DAY,
+    EMPTY
+}
