+++ app/src/main/java/kr/co/nottodo/view/calendar/monthly/model/DateType.kt
@@ -1,0 +1,7 @@
+package kr.co.nottodo.view.calendar.monthly.model
+
+enum class DateType {
+    WEEKDAY,
+    DISABLED,
+    WEEKEND
+}
