+++ app/src/main/java/kr/co/nottodo/view/calendar/monthly/monthlycalendar/MonthlyCalendar.kt
@@ -1,0 +1,286 @@
+package kr.co.nottodo.view.calendar.monthly.monthlycalendar
+
+import android.content.Context
+import android.graphics.Color
+import android.util.AttributeSet
+import android.view.Gravity
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.LinearLayout
+import android.widget.TextView
+import androidx.core.content.ContextCompat
+import androidx.core.view.ViewCompat
+import androidx.recyclerview.widget.GridLayoutManager
+import kr.co.nottodo.R
+import kr.co.nottodo.databinding.ViewCalendarWeekDescriptionBinding
+import kr.co.nottodo.util.extension.dpToPx
+import kr.co.nottodo.view.NoRippleRecyclerView
+import kr.co.nottodo.view.calendar.monthly.monthlycalendar.adapter.MonthlyCalendarDayAdapter
+import kr.co.nottodo.view.calendar.monthly.model.DAY_COLUMN_COUNT
+import kr.co.nottodo.view.calendar.monthly.model.DateType
+import kr.co.nottodo.view.calendar.monthly.model.CalendarDay
+import kr.co.nottodo.view.calendar.monthly.model.TOTAL_COLUMN_COUNT
+import kr.co.nottodo.view.calendar.util.isWeekend
+import kr.co.nottodo.view.calendar.util.toPrettyDateString
+import kr.co.nottodo.view.calendar.util.toPrettyMonthString
+import java.util.*
+import java.util.Calendar.*
+
+/**
+ * created by ssong-develop on 2022.12.31
+ *
+ * No Swipe Effection MonthlyCalendar
+ *
+ * 성취 뷰에서 사용합니다.
+ */
+class MonthlyCalendar @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyle: Int = 0
+) : LinearLayout(context, attrs, defStyle) {
+
+    private val timeZone = TimeZone.getDefault()
+    private val locale = Locale.KOREA
+    private val monthlyCalendarDayAdapter = MonthlyCalendarDayAdapter()
+    private val calendar = Calendar.getInstance(timeZone, locale)
+    private var calendarDataList: List<CalendarDay> = listOf()
+    private var currentDate = calendar.toPrettyMonthString(locale = locale)
+        set(value) {
+            field = value
+            updateCurrentDateTextView()
+        }
+
+    private val currentDateTextView = TextView(context, null, R.style.M12).apply {
+        id = ViewCompat.generateViewId()
+        text = currentDate
+        layoutParams =
+            LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT)
+        setTextColor(ContextCompat.getColor(context, R.color.gray_1_626068))
+    }
+
+    private val calendarHeaderLinearLayout = LinearLayout(context).apply {
+        id = ViewCompat.generateViewId()
+        orientation = HORIZONTAL
+        gravity = Gravity.CENTER_VERTICAL
+        setPadding(context.dpToPx(30), context.dpToPx(24), 0, context.dpToPx(10))
+        addView(
+            ImageView(this.context).apply {
+                setImageDrawable(
+                    ContextCompat.getDrawable(
+                        this.context,
+                        R.drawable.ic_arrow_left_gray4
+                    )
+                )
+                setOnClickListener {
+                    calendar.add(Calendar.MONTH, -1)
+                    currentDate = calendar.toPrettyMonthString(locale = locale)
+                    initCalendarData()
+                }
+            }
+        )
+
+        addView(currentDateTextView)
+
+        addView(
+            ImageView(this.context).apply {
+                setImageDrawable(
+                    ContextCompat.getDrawable(
+                        this.context,
+                        R.drawable.ic_arrow_right_gray4
+                    )
+                )
+                setOnClickListener {
+                    calendar.add(Calendar.MONTH, 1)
+                    currentDate = calendar.toPrettyMonthString(locale = locale)
+                    initCalendarData()
+                }
+            }
+        )
+    }
+
+    private val calendarWeekDescriptionView = ViewCalendarWeekDescriptionBinding.inflate(
+        LayoutInflater.from(context), this, false
+    )
+
+    private val monthRecyclerView = NoRippleRecyclerView(context).apply {
+        id = ViewCompat.generateViewId()
+        adapter = monthlyCalendarDayAdapter
+        layoutManager = GridLayoutManager(context, TOTAL_COLUMN_COUNT).apply {
+            spanSizeLookup = object : GridLayoutManager.SpanSizeLookup() {
+                override fun getSpanSize(position: Int): Int {
+                    return DAY_COLUMN_COUNT
+                }
+            }
+        }
+        layoutParams =
+            LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT)
+    }
+
+    init {
+        if (attrs != null) {
+            getStyleableAttrs(attrs)
+        }
+        layoutParams = LayoutParams(LayoutParams.MATCH_PARENT, LayoutParams.WRAP_CONTENT)
+        orientation = LinearLayout.VERTICAL
+
+        addView(calendarHeaderLinearLayout)
+        addView(calendarWeekDescriptionView.root)
+        addView(monthRecyclerView)
+        initBackgroundColor()
+        initializeNotToDoMonthCalendar()
+    }
+
+    private fun updateCurrentDateTextView() {
+        currentDateTextView.text = currentDate
+    }
+
+    private fun initializeNotToDoMonthCalendar() {
+        initCalendarData()
+    }
+
+    private fun initCalendarData() {
+        calendarDataList = buildCalendarData()
+        monthlyCalendarDayAdapter.submitList(calendarDataList)
+    }
+
+    private fun buildCalendarData(): List<CalendarDay> {
+        // 현재 달력이 보여주고 있는 Calendar instance의 복사본
+        val proxyCalendar = Calendar.getInstance().apply {
+            set(MONTH, calendar.get(MONTH))
+            set(DAY_OF_MONTH, calendar.get(DAY_OF_MONTH))
+            set(YEAR, calendar.get(YEAR))
+        }
+
+        val totalDayInMonth = calendar.getActualMaximum(Calendar.DAY_OF_MONTH)
+        val calendarDayList = mutableListOf<CalendarDay>()
+        (1..totalDayInMonth).forEach { day ->
+            proxyCalendar.set(Calendar.DAY_OF_MONTH, day)
+            val dayOfWeek = proxyCalendar.get(Calendar.DAY_OF_WEEK)
+            val dateType = if (proxyCalendar.isWeekend()) {
+                DateType.WEEKEND
+            } else {
+                DateType.WEEKDAY
+            }
+
+            when (day) {
+                1 -> {
+                    calendarDayList.addAll(
+                        createStartEmptyView(
+                            dayOfWeek,
+                            proxyCalendar
+                        )
+                    )
+                    calendarDayList.add(
+                        CalendarDay.Day(
+                            day.toString(),
+                            proxyCalendar.toPrettyDateString(),
+                            proxyCalendar.time,
+                            state = dateType
+                        )
+                    )
+                }
+                totalDayInMonth -> {
+                    calendarDayList.add(
+                        CalendarDay.Day(
+                            day.toString(),
+                            proxyCalendar.toPrettyDateString(),
+                            proxyCalendar.time,
+                            state = dateType
+                        )
+                    )
+                    calendarDayList.addAll(
+                        createEndEmptyView(
+                            dayOfWeek,
+                            proxyCalendar
+                        )
+                    )
+                }
+                else -> {
+                    calendarDayList.add(
+                        CalendarDay.Day(
+                            day.toString(),
+                            proxyCalendar.toPrettyDateString(),
+                            proxyCalendar.time,
+                            state = dateType
+                        )
+                    )
+                }
+            }
+        }
+        return calendarDayList
+    }
+
+    private fun createEndEmptyView(
+        dayOfWeek: Int,
+        proxyCalendar: Calendar
+    ): List<CalendarDay.Empty> {
+        val nextCalendar = Calendar.getInstance().apply {
+            set(MONTH, proxyCalendar.get(MONTH))
+            set(DAY_OF_MONTH, proxyCalendar.get(DAY_OF_MONTH))
+            set(YEAR, proxyCalendar.get(YEAR))
+        }.also {
+            it.add(MONTH, 1)
+        }
+        var totalDayInNextMonth = nextCalendar.getActualMinimum(DAY_OF_MONTH)
+        val numberOfEmptyView = when (dayOfWeek) {
+            Calendar.SUNDAY -> 6
+            Calendar.MONDAY -> 5
+            Calendar.TUESDAY -> 4
+            Calendar.WEDNESDAY -> 3
+            Calendar.THURSDAY -> 2
+            Calendar.FRIDAY -> 1
+            else -> 0
+        }
+
+        val listEmpty = mutableListOf<CalendarDay.Empty>()
+        repeat((0 until numberOfEmptyView).count()) {
+            listEmpty.add(
+                CalendarDay.Empty(
+                    totalDayInNextMonth++.toString()
+                )
+            )
+        }
+        return listEmpty
+    }
+
+    private fun createStartEmptyView(
+        dayOfWeek: Int,
+        proxyCalendar: Calendar
+    ): List<CalendarDay.Empty> {
+        val previousCalendar = Calendar.getInstance().apply {
+            set(MONTH, proxyCalendar.get(MONTH))
+            set(DAY_OF_MONTH, proxyCalendar.get(DAY_OF_MONTH))
+            set(YEAR, proxyCalendar.get(YEAR))
+        }.also {
+            it.add(MONTH, -1)
+        }
+        val numberOfEmptyView = when (dayOfWeek) {
+            Calendar.MONDAY -> 1
+            Calendar.TUESDAY -> 2
+            Calendar.WEDNESDAY -> 3
+            Calendar.THURSDAY -> 4
+            Calendar.FRIDAY -> 5
+            Calendar.SATURDAY -> 6
+            else -> 0
+        }
+        var startDayInPreviousMonth = previousCalendar.getActualMaximum(DAY_OF_MONTH) - numberOfEmptyView + 1
+        val listEmpty = mutableListOf<CalendarDay.Empty>()
+        repeat((0 until numberOfEmptyView).count()) {
+            listEmpty.add(
+                CalendarDay.Empty(
+                    startDayInPreviousMonth++.toString()
+                )
+            )
+        }
+        return listEmpty
+    }
+
+    private fun initBackgroundColor() {
+        setBackgroundColor(Color.parseColor("#ffffff"))
+    }
+
+    private fun getStyleableAttrs(attrs: AttributeSet) {
+
+    }
+}
