+++ app/src/main/java/kr/co/nottodo/view/calendar/monthly/monthlycalendar/viewholder/MonthlyCalendarViewHolder.kt
@@ -1,0 +1,9 @@
+package kr.co.nottodo.view.calendar.monthly.monthlycalendar.viewholder
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+import kr.co.nottodo.view.calendar.monthly.model.CalendarDay
+
+abstract class MonthlyCalendarViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+    abstract fun onBind(data: CalendarDay)
+}
