+++ app/src/main/java/kr/co/nottodo/view/calendar/monthly/monthlycalendarpicker/listener/MonthlyCalendarPickerClickListener.kt
@@ -1,0 +1,8 @@
+package kr.co.nottodo.view.calendar.monthly.monthlycalendarpicker.listener
+
+import android.view.View
+import java.util.Date
+
+fun interface MonthlyCalendarPickerClickListener {
+    fun onDayClick(view: View, date: Date)
+}
