+++ app/src/main/java/kr/co/nottodo/view/calendar/util/CalendarExtension.kt
@@ -1,0 +1,21 @@
+package kr.co.nottodo.view.calendar.util
+
+import java.util.*
+
+fun Date.isTheSameDay(comparedDate: Date): Boolean {
+    val calendar = Calendar.getInstance()
+    calendar.withTime(this)
+    val comparedCalendarDate = Calendar.getInstance()
+    comparedCalendarDate.withTime(comparedDate)
+    return calendar.get(Calendar.DAY_OF_YEAR) == comparedCalendarDate.get(Calendar.DAY_OF_YEAR) &&
+            calendar.get(Calendar.YEAR) == comparedCalendarDate.get(Calendar.YEAR)
+}
+
+fun Calendar.withTime(date: Date) {
+    clear()
+    time = date
+    set(Calendar.HOUR_OF_DAY, 0)
+    set(Calendar.MINUTE, 0)
+    set(Calendar.SECOND, 0)
+    set(Calendar.MILLISECOND, 0)
+}
