--- app/src/main/java/kr/co/nottodo/view/monthlycalendar/model/CalendarType.kt
@@ -1,7 +1,0 @@
-package kr.co.nottodo.view.monthlycalendar.model
-
-enum class CalendarType {
-    WEEK,
-    DAY,
-    EMPTY
-}
