--- app/src/main/java/kr/co/nottodo/view/monthlycalendar/model/DateType.kt
@@ -1,7 +1,0 @@
-package kr.co.nottodo.view.monthlycalendar.model
-
-enum class DateType {
-    WEEKDAY,
-    DISABLED,
-    WEEKEND
-}
