--- app/src/main/java/kr/co/nottodo/view/monthlycalendarpicker/listener/MonthlyCalendarPickerClickListener.kt
@@ -1,8 +1,0 @@
-package kr.co.nottodo.view.monthlycalendarpicker.listener
-
-import android.view.View
-import java.util.Date
-
-fun interface MonthlyCalendarPickerClickListener {
-    fun onDayClick(view: View, date: Date)
-}
