--- app/src/main/java/kr/co/nottodo/view/weeklycalendar/NotToDoWeeklyAdapter.kt
@@ -1,4 +1,0 @@
-package kr.co.nottodo.view.weeklycalendar
-
-class NotToDoWeeklyAdapter {
-}
