--- app/src/main/java/kr/co/nottodo/view/weeklycalendar/NotToDoWeeklyViewHolder.kt
@@ -1,4 +1,0 @@
-package kr.co.nottodo.view.weeklycalendar
-
-class NotToDoWeeklyViewHolder {
-}
