+++ app/src/main/res/layout/view_monthly_calendar_empty.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="empty_day_item"
-            type="kr.co.nottodo.view.monthlycalendar.model.CalendarDay.Empty" />
+            type="kr.co.nottodo.view.calendar.monthly.model.CalendarDay" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
