+++ app/src/main/res/layout/view_no_use_calendar_month.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="month_item"
-            type="kr.co.nottodo.view.monthlycalendar.model.CalendarMonth" />
+            type="kr.co.nottodo.view.calendar.monthly.model.CalendarMonth" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
